requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7 